Refactor order picking logic to optimize inventory allocation
Revised `PickingItem` and `OrderService` to improve picking strategies by handling cases where single locations can't fulfill orders and ensuring logical sorting of picking routes. Updated `productId` type to `long` for consistency across the system. Added a `toString` method to `ProductDTO` for better debugging and logging.

+++ src/main/java/be/vdab/scrumjava202504/orders/PickingItem.java
@@ -5,10 +5,10 @@ public class PickingItem {
     private int position;
     private String name;
     private int quantityOrdered;
-    private int productId;
+    private long productId;
     private boolean picked = false;
 
-    public PickingItem(String shelf, int position, String name, int quantity, int productId) {
+    public PickingItem(String shelf, int position, String name, int quantity, long productId) {
         this.shelf = shelf;
         this.position = position;
         this.name = name;
